feat-created the request interface for concrete methods

+++ src/main/java/dev/sara/request/RequestController.java
@@ -38,7 +38,7 @@ public class RequestController {
     }
 
     @GetMapping("/{id}")
-    public String show(@PathVariable("id") int id) {
+    public String show(@PathVariable("id") Long id) {
         return new String();
     }
 }
